+++ crates/autoeq-cli/src/lib.rs
@@ -7,24 +7,4 @@ pub mod benchmark;
 pub mod download;
 
 pub use autoeq_workflow::cli::*;
-
-/// Conditional debug output used by the historical AutoEQ CLI.
-#[macro_export]
-macro_rules! qa_println {
-    ($fmt:literal) => {
-        log::debug!($fmt);
-    };
-    ($fmt:literal, $($arg:expr),* $(,)?) => {
-        log::debug!($fmt, $($arg),*);
-    };
-    ($args:expr, $fmt:literal) => {
-        if $args.qa.is_none() {
-            log::debug!($fmt);
-        }
-    };
-    ($args:expr, $fmt:literal, $($arg:expr),* $(,)?) => {
-        if $args.qa.is_none() {
-            log::debug!($fmt, $($arg),*);
-        }
-    };
-}
+pub use autoeq_workflow::qa_println;